feat:create RequestRepository and Mapper

+++ ewm-service/src/main/java/ru/practicum/model/dto/ParticipationRequestDto.java
@@ -15,6 +15,6 @@ public class ParticipationRequestDto {
     private LocalDateTime created;
     private Long event;
     private Long id;
-    private Integer requester;
+    private Long requester;
     private State status;
 }
